Add material colorscheme
Add color scheme Material and default to the "lighter" style.  Useful when working under the sun. One issue with this color scheme is that the underlying tmux disabled window color is not showing. The onedark theme has a transparency setting that allows this, but I haven't found a way to make this happen and still have the "lighter" style display correctly.